Manual data normalisation and author verification_V2
Due to a bug in the data_normalizer where it fails to properly normalize the raw data from GitHub after execution, the data is being manually normalized for now. Additionally, the authors' names have been manually reviewed and verified.

This version enhances the readability of ymal files to make them easier to maintain.